+++ src/pages/TitlePage/TitlePage.tsx
@@ -1,7 +1,8 @@
 import Shell from '../../components/layout/Shell/Shell';
 import TitleEntry from '../entry/TitleEntry/TitleEntry';
+import { PAGE_TITLE_ADD } from '@/utils/constants';
 
-const PAGE = 0;
+const PAGE = PAGE_TITLE_ADD;
 
 export default function TitlePage() {
   return (